refactor: reorganize API client tests into modular structure
- Replaced duplicate test files (test_client.py, test_client_comprehensive.py)
- Created organized test directory: tests/unit/api_clients/client/
- Split tests into logical modules:
  * test_initialization.py - client setup and configuration
  * test_endpoints.py - API endpoint methods and data retrieval
  * test_async.py - async functionality and context management
  * test_error_handling.py - error scenarios and exception handling
- Eliminated test duplication and improved maintainability
- Enhanced test organization for better coverage analysis
- Maintained comprehensive test coverage while improving structure

Addresses GitHub Issue #26: API Client Coverage Optimization
