+++ tests/unit/api_clients/client/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for AlphaVantageClient module."""
